seperate auth and spam helper
