+++ cli/spam.go
@@ -58,11 +58,11 @@ func (h *Handler) BuildSpammer(sh loadgen.SpamHelper, defaults bool) (*loadgen.S
 			balance,
 			1.01,
 			2.7,
+			100000, // tx per second
+			15000, // min tx per second
+			1000, // tx per second step
 			10,
-			10,
-			10,
-			10,
-			10,
+			10000000, // num accounts 
 		), nil
 	}
 	// Collect parameters
