+++ examples/morpheusvm/auth/keys.go
@@ -1,0 +1,106 @@
+package auth
+
+import (
+	"errors"
+	"fmt"
+
+	"github.com/ava-labs/hypersdk/auth"
+	"github.com/ava-labs/hypersdk/crypto/bls"
+	"github.com/ava-labs/hypersdk/crypto/ed25519"
+	"github.com/ava-labs/hypersdk/crypto/secp256r1"
+	"github.com/ava-labs/hypersdk/utils"
+)
+
+const (
+	ed25519Key   = "ed25519"
+	secp256r1Key = "secp256r1"
+	blsKey       = "bls"
+)
+
+var (
+	ErrInvalidKeyType    = errors.New("invalid key type")
+)
+
+func CheckKeyType(k string) error {
+	switch k {
+	case ed25519Key, secp256r1Key, blsKey:
+		return nil
+	default:
+		return fmt.Errorf("%w: %s", ErrInvalidKeyType, k)
+	}
+}
+
+func GeneratePrivateKey(k string) (*auth.PrivateKey, error) {
+	switch k {
+	case ed25519Key:
+		p, err := ed25519.GeneratePrivateKey()
+		if err != nil {
+			return nil, err
+		}
+		return &auth.PrivateKey{
+			Address: auth.NewED25519Address(p.PublicKey()),
+			Bytes:   p[:],
+		}, nil
+	case secp256r1Key:
+		p, err := secp256r1.GeneratePrivateKey()
+		if err != nil {
+			return nil, err
+		}
+		return &auth.PrivateKey{
+			Address: auth.NewSECP256R1Address(p.PublicKey()),
+			Bytes:   p[:],
+		}, nil
+	case blsKey:
+		p, err := bls.GeneratePrivateKey()
+		if err != nil {
+			return nil, err
+		}
+		return &auth.PrivateKey{
+			Address: auth.NewBLSAddress(bls.PublicFromPrivateKey(p)),
+			Bytes:   bls.PrivateKeyToBytes(p),
+		}, nil
+	default:
+		return nil, ErrInvalidKeyType
+	}
+}
+
+func LoadPrivateKey(k string, path string) (*auth.PrivateKey, error) {
+	switch k {
+	case ed25519Key:
+		p, err := utils.LoadBytes(path, ed25519.PrivateKeyLen)
+		if err != nil {
+			return nil, err
+		}
+		pk := ed25519.PrivateKey(p)
+		return &auth.PrivateKey{
+			Address: auth.NewED25519Address(pk.PublicKey()),
+			Bytes:   p,
+		}, nil
+	case secp256r1Key:
+		p, err := utils.LoadBytes(path, secp256r1.PrivateKeyLen)
+		if err != nil {
+			return nil, err
+		}
+		pk := secp256r1.PrivateKey(p)
+		return &auth.PrivateKey{
+			Address: auth.NewSECP256R1Address(pk.PublicKey()),
+			Bytes:   p,
+		}, nil
+	case blsKey:
+		p, err := utils.LoadBytes(path, bls.PrivateKeyLen)
+		if err != nil {
+			return nil, err
+		}
+
+		privKey, err := bls.PrivateKeyFromBytes(p)
+		if err != nil {
+			return nil, err
+		}
+		return &auth.PrivateKey{
+			Address: auth.NewBLSAddress(bls.PublicFromPrivateKey(privKey)),
+			Bytes:   p,
+		}, nil
+	default:
+		return nil, ErrInvalidKeyType
+	}
+}
